+++ src/egon/data/datasets/power_plants/assign_weather_data.py
@@ -2,7 +2,6 @@ import geopandas as gpd
 import pandas as pd
 
 from egon.data import db
-from egon.data.datasets import Dataset
 import egon.data.config
 import egon.data.datasets.power_plants.__init__ as init_pp
 
@@ -19,9 +18,11 @@ def find_bus_id(power_plants, cfg):
     power_plants = power_plants[~power_plants.bus_id.isna()]
 
     power_plants_no_busId = power_plants_no_busId.drop(columns="bus_id")
-    
+
     if len(power_plants_no_busId) > 0:
-        power_plants_no_busId = init_pp.assign_bus_id(power_plants_no_busId, cfg)
+        power_plants_no_busId = init_pp.assign_bus_id(
+            power_plants_no_busId, cfg
+        )
 
     power_plants = power_plants.append(power_plants_no_busId)
 
@@ -107,10 +108,10 @@ def write_power_plants_table(power_plants, cfg, con):
 
     # delete weather dependent power_plants from supply.egon_power_plants
     db.execute_sql(
-        f""" 
+        f"""
     DELETE FROM {cfg['sources']['power_plants']['schema']}.
-    {cfg['sources']['power_plants']['table']} 
-    WHERE carrier IN ('wind_onshore', 'solar') 
+    {cfg['sources']['power_plants']['table']}
+    WHERE carrier IN ('wind_onshore', 'solar')
     """
     )
 
@@ -132,7 +133,7 @@ def write_power_plants_table(power_plants, cfg, con):
     """
     max_id = pd.read_sql(sql, con)
     max_id = max_id["max"].iat[0]
-    if max_id == None:
+    if max_id is None:
         ini_id = 1
     else:
         ini_id = int(max_id + 1)
@@ -152,4 +153,3 @@ def write_power_plants_table(power_plants, cfg, con):
     )
 
     return "Bus_id and Weather_id were updated succesfully"
-